requirements.txt
Fabric==1.8.2
Flask==0.10.1
Flask-Login==0.2.9
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.4
Wand==0.3.5
facebook-sdk==0.4.0
psycopg2==2.5.2
Flask-Assets==0.9
jsmin==2.0.9
cssmin==0.2.0
celery==3.1.10
